fix views structure
